* manage step() from meshes better * remove some warnings * whitespace fixes * lofted-polygon-segment intersection * make 3D mesher more robust * initialisation fixes * viscoelasticity getCopy() fixes
git-svn-id: svn+ssh://c4science.ch/source/amie/trunk@2913 64fb8e13-074e-4207-b257-2d34fb7690c6

+++ examples/main_asr_simple.cpp
@@ -98,7 +98,7 @@ void step()
 {
 	featureTree->setMaxIterationsPerStep(2000) ;
 	int nsteps = 1;
-	for(size_t i = 0 ; i < nsteps ; i++)
+	for(int i = 0 ; i < nsteps ; i++)
 	{
 
 		bool go_on = featureTree->step() ;
